modify dialog style

+++ src/views/customers/index.vue
@@ -203,11 +203,11 @@
         :model="selectedCustomer"
         label-position="top"
         label-width="100px"
-        class="single-item"
       >
         <el-form-item
           :label="$t('customersView.title')"
           prop="title"
+          class="single"
         >
           <el-input
             v-model="selectedCustomer.title"
@@ -217,12 +217,14 @@
         <el-form-item
           :label="$t('customersView.authorizedPersonName')"
           prop="authorizedPerson"
+          class="single"
         >
           <el-input v-model="selectedCustomer.authorized_person_name" />
         </el-form-item>
         <el-form-item
           :label="$t('customersView.phoneNumber')"
           prop="phone"
+          class="single"
         >
           <el-input
             v-model="selectedCustomer.phone_number"
@@ -447,6 +449,11 @@ export default class extends Vue {
 
   private deleteCustomer(customer: ICustomer) {
     this.loading = true
+
+    if (customer.id === null) {
+      console.error('customer id cannot be null')
+      return
+    }
 
     service.deleteCustomer(customer.id)
       .then(
